requirements.txt
# Framework Django
Django==4.2.27
django-filter==21.1
djangorestframework==3.15.1

# Swagger
drf-yasg==1.21.14

# Static files
whitenoise==6.5.0

# ASGI
asgiref==3.8.1
uvicorn[standard]

# Base de datos PostgreSQL
psycopg2-binary==2.9.9
psycopg[pool]==3.2.3

# Authentication
djangorestframework-simplejwt==5.5.1

# Async DRF
adrf==0.1.7

# Caching
redis==5.2.1
django-redis==5.4.0
